display: oplus: get vblank when switching between ON and LP1
Whenever dimlayer status is changed via /dev/oplus_display, the function
for that gets a vblank, along with setting up a vblank put to occur
after 5 seconds. While I don't entirely understand the idea behind this,
it seems to prevent situations where High Brightness Mode gets enabled
before dimlayer is set up.

With that in mind, whenever we switch between ON and LP1 while
dimlayer_hbm is enabled, the dimlayer is still set up, but its opacity
is changed depending on the mode. Without getting a vblank, the dimlayer
transparency doesn't change in time, resulting in fun situations where
the user gets blasted with full HBM brightness for a frame or two when
switching the panel back to ON mode.

Let's just do the same vblank get/put when switching between ON and LP1
modes.

Change-Id: I73288152f7048d94b3081cfd536df37a17577950